[update] 소유주 확인 로직 추가, userId 스프링 시큐리티에서 가져오게 변경

+++ src/main/java/com/dahhong/whoami/page/adapter/in/PageController.java
@@ -1,5 +1,7 @@
 package com.dahhong.whoami.page.adapter.in;
 
+import com.dahhong.whoami.auth.domain.entity.UserDetailsImpl;
+import com.dahhong.whoami.global.exception.customException.AuthorizationFailureException;
 import com.dahhong.whoami.global.response.ApiResponse;
 import com.dahhong.whoami.page.adapter.in.dto.PageRequestDto;
 import com.dahhong.whoami.page.adapter.in.dto.CreatePageResponseDto;
@@ -10,8 +12,13 @@ import com.dahhong.whoami.page.application.port.in.GetPageUseCase;
 import com.dahhong.whoami.page.application.port.in.UpdatePageUseCase;
 import com.dahhong.whoami.page.application.service.DeletePageService;
 import com.dahhong.whoami.page.domain.entity.Page;
+import jakarta.validation.Valid;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.ResponseEntity;
+import org.springframework.security.core.Authentication;
+import org.springframework.security.core.annotation.AuthenticationPrincipal;
+import org.springframework.security.core.context.SecurityContextHolder;
+import org.springframework.security.core.userdetails.UserDetails;
 import org.springframework.web.bind.annotation.*;
 
 import java.util.List;
@@ -42,19 +49,28 @@ public class PageController implements PageControllerSwagger {
 	}
 
 	@PostMapping("/create")
-	public ResponseEntity<?> createPage(@RequestBody PageRequestDto pageRequest) {
-		Page createdPage = createPageUseCase.createPage(pageRequest.getUserId(), pageRequest.getTitle());
+	public ResponseEntity<?> createPage(@Valid @RequestBody PageRequestDto pageRequest, @AuthenticationPrincipal String userId) {
+		/*if (userId == null) {
+			throw new AuthorizationFailureException("인증 정보가 없습니다.", null);
+		}*/
+		Page createdPage = createPageUseCase.createPage(userId, pageRequest.getTitle());
 		return ResponseEntity.ok(ApiResponse.success(new CreatePageResponseDto("성공적으로 페이지를 생성했습니다", createdPage.getId())));
 	}
 
 	@PutMapping("/{id}")
-	public ResponseEntity<?> updatePage(@PathVariable Long id, @RequestBody PageRequestDto pageRequest) {
+	public ResponseEntity<?> updatePage(@PathVariable Long id, @Valid @RequestBody PageRequestDto pageRequest, @AuthenticationPrincipal String userId) {
+		if (getPageUseCase.checkPageOwnership(id, userId)) {
+			throw new AuthorizationFailureException("인증 정보가 일치하지 않습니다.", null);
+		}
 		updatePageUseCase.updatePage(id, pageRequest);
 		return ResponseEntity.ok(ApiResponse.success());
 	}
 
 	@DeleteMapping("/{id}")
-	public ResponseEntity<?> deletePage(@PathVariable Long id) {
+	public ResponseEntity<?> deletePage(@PathVariable Long id, @AuthenticationPrincipal String userId) {
+		if (getPageUseCase.checkPageOwnership(id, userId)) {
+			throw new AuthorizationFailureException("인증 정보가 일치하지 않습니다.", null);
+		}
 		deletePageUseCase.deletePage(id);
 		return ResponseEntity.ok(ApiResponse.success());
 	}
